# youthchinaopenapi.yml
openapi: 3.0.0
info:
  version: 1.5.3
  title: YouthChina
  license:
    name: MIT
servers:
  - url: 'http://youthchina.io/api/v1'
paths:
...
  '/questions':
    post:
      tags:
        - Question
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionRequest'
      responses:
        '201':
          description: create question
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: '#/components/schemas/Question'
                  status:
                    $ref: '#/components/schemas/Status'
        '400':
          description: cannot create question due to user's error
    get:
      tags:
        - Question
      parameters:
        - name: company
          in: query
          description: company name
          schema:
            type: string
        - name: job
          in: query 
          description: job name
          schema:
            type: string
      responses:
        '200':
          description: search job
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    properties:
                      questions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Question'                     
                  status:
                    $ref: '#/components/schemas/Status'
        '400':
          description: cannot create question due to user's error
  '/questions/{id}':
    get:
      tags:
        - Question
      parameters:
        - name: id
          required: true
          in: path
          description: question id
          schema:
            type: integer
      responses:
        '200':
          description: return target question
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: '#/components/schemas/Question'
                  status:
                    $ref: '#/components/schemas/Status'
        '404':
          description: target question does not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStatus'
    delete:
      tags:
        - Question
      parameters:
        - name: id
          required: true
          in: path
          description: question id
          schema:
            type: integer
      responses:
        '204':
          description: deleted successed.
        '404':
          description: target question does not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStatus'
    put:
      tags:
        - Question
      parameters:
        - name: id
          required: true
          in: path
          description: id of the question
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionRequest'
      responses:
        '204':
          description: updated success
  '/questions/{id}/answers':
    get:
      tags:
        - Question
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer

      responses:
        '200':
          description: the answer list of target question
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: object
                    properties:
                      answers:
                        type: array
                        items:
                          $ref: '#/components/schemas/AnswerBasic'
                  status:
                    $ref: '#/components/schemas/Status'
    post:
      tags:
        - Answer
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
              - id
              - body
              properties:
                id:
                  type: integer
                body:
                  $ref: '#/components/schemas/RichText'
                is_anonymous:
                  type: boolean
                  default: false
                create_at:
                  type: integer
                
      responses:
        '201':
          description: target answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: '#/components/schemas/Answer'
                  status:
                    $ref: '#/components/schemas/Status'
  '/questions/{questionId}/invite/{userId}':
    put:
      tags:
        - Question
      parameters:
        - name: questionId
          required: true
          in: path
          schema:
            type: integer
        - name: userId
          required: true
          in: path
          schema:
            type: integer
      responses:
        '201':
          description: invitation send successfully.
...
  '/questions/{id}/attention':
    put:
      tags:
        - Question
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses: 
        '201':
          description: upvote successfullly
  '/answers/{id}':
    get:
      tags:
        - Answer
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        '200':
          description: target answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: '#/components/schemas/Answer'
                  status:
                    $ref: '#/components/schemas/Status'
    put:
      tags:
        - Answer
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerRequest'
      responses:
        '200':
          description: target answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: '#/components/schemas/Answer'
                  status:
                    $ref: '#/components/schemas/Status'
    delete:
      tags:
        - Answer
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: integer
      responses:
        '204':
          description: success
        '403':
          description: do not have authority to delete answer
...
components:
  schemas:
...
    Status:
      type: object
      properties:
        code:
          type: integer
          format: int64
        reason:
          type: string
    ErrorStatus:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/Status'
...
    User:
      type: object
      required:
        - id
        - username
        - email
        - phonenumber
        - register_date
        - real_name
        - gender
        - nation
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        phonenumber:
          type: string
        register_date:
          type: string
        real_name:
          type: string
        gender:
          type: string
          enum:
            - male
            - female
        nation:
          type: string
        avatar_url:
          type: string
        role:
          type: string
        age:
          type: integer
...
    AnswerRequest:
      type: object
      required:
        - body
        - is_anonymous
      properties:
        body:
          $ref: '#/components/schemas/RichText'
        is_anonymous:
          type: boolean
          default: false
    AnswerBasic:
      allOf:
        - $ref: '#/components/schemas/AnswerRequest'
        - properties:
            id:
              type: integer
              description: answer id
            creator:
              $ref: '#/components/schemas/User'
            modified_at:
              type: string
            create_at:
              type: string
    QuestionRequest:
      type: object
      required:
        - title
        - body
        - is_anonymous
      properties:
        title:
          type: string
          description: question title
        body:
          $ref: '#/components/schemas/RichText'
        is_anonymous:
          type: boolean
          default: flase
        rela_type:
          type: integer
          description: 1 for no rela, 2 for company, 3 for job
        rela_id:
          type: integer
    QuestionBasic:
      allOf:
        - $ref: '#/components/schemas/QuestionRequest'
        - properties:
            id:
              type: integer
              description: question id
            create_at:
              type: string
            modified_at:
              type: string
            creator:
              $ref: '#/components/schemas/User'
    Question:
      allOf:
        - $ref: '#/components/schemas/QuestionBasic'
        - type: object
          properties:
            answers:
              type: array
              items:
                $ref: '#/components/schemas/AnswerBasic'
    Answer:
      allOf:
        - $ref: '#/components/schemas/AnswerBasic'
        - type: object
          properties:
            question:
              $ref: '#/components/schemas/QuestionBasic'
    RichText:
      type: object
      required:
        - braftEditorRaw
        - previewText
        - resourceList
      properties:
        braftEditorRaw:
          type: object
        previewText:
          type: object
        resourceList:
          type: array
          items: 
            type: string
...